Change kpay amount int to float

+++ src/KbzPayPwa.php
@@ -8,7 +8,7 @@ class KbzPayPwa extends KbzPay
     /**
      * @throws Exception
      */
-    public function getPaymentScreenUrl(string $orderId, int $amount,string $nonceStr, string $backendResultUrl): string
+    public function getPaymentScreenUrl(string $orderId, float $amount,string $nonceStr, string $backendResultUrl): string
     {
         return $this->preCreate("PWAAPP", $orderId, $amount, $nonceStr, $backendResultUrl);
     }
